+++ src/Application/Command/User/SignIn/SignInHandler.php
@@ -5,24 +5,21 @@ declare(strict_types=1);
 namespace App\Application\Command\User\SignIn;
 
 use App\Application\Command\CommandHandlerInterface;
-use App\Domain\User\Exception\Credentials\InvalidCredentialsException;
+use App\Application\Service\ObtainUuidFromEmail;
 use App\Domain\User\Repository\UserEventStoreInterface;
-use App\Domain\User\Service\UserFinder;
-use App\Domain\User\ValueObj\Email;
-use Ramsey\Uuid\UuidInterface;
 
 class SignInHandler implements CommandHandlerInterface
 {
-    /** @var UserFinder */
-    private $userFinder;
+    /** @var ObtainUuidFromEmail */
+    private $obtainUuid;
 
     /** @var UserEventStoreInterface */
     private $eventStore;
 
 
-    public function __construct(UserFinder $finderByEmail, UserEventStoreInterface $userEventStore)
+    public function __construct(ObtainUuidFromEmail $obtainUuidFromEmail, UserEventStoreInterface $userEventStore)
     {
-        $this->userFinder = $finderByEmail;
+        $this->obtainUuid = $obtainUuidFromEmail;
         $this->eventStore = $userEventStore;
     }
 
@@ -33,27 +30,11 @@ class SignInHandler implements CommandHandlerInterface
      */
     public function __invoke(SignInCommand $command): void
     {
-        $uuid = $this->getUuidFromEmail($command->email);
+        $uuid = $this->obtainUuid->get($command->email);
         $user = $this->eventStore->get($uuid);
 
         $user->signIn($command->plainPassword);
 
         $this->eventStore->store($user);
-    }
-
-
-    /**
-     * @param  Email         $email
-     * @return UuidInterface
-     */
-    private function getUuidFromEmail(Email $email): UuidInterface
-    {
-        $userView = $this->userFinder->findByEmail($email);
-
-        if (null === $userView) {
-            throw new InvalidCredentialsException('Invalid credentials entered.');
-        }
-
-        return $userView->getUuid();
     }
 }